Abort if passthrough is specified for Metanorma XML, but XML is not compliant: https://github.com/metanorma/metanorma/issues/437
