+++ lib/metanorma/standoc/validate.rb
@@ -2,8 +2,8 @@ require "metanorma/standoc/utils"
 require_relative "validate_section"
 require_relative "validate_table"
 require_relative "validate_term"
+require_relative "validate_schema"
 require "nokogiri"
-require "jing"
 require "iev"
 require "pngcheck"
 
@@ -96,95 +96,6 @@ module Metanorma
         err2 = "There is a crossreference to an instance of #{inner.name} " \
                "nested within #{outer.name}: #{i.to_xml}"
         @log.add("Style", i, err2)
-      end
-
-      def schema_validate(doc, schema)
-        Tempfile.open(["tmp", ".xml"], encoding: "UTF-8") do |f|
-          schema_validate1(f, doc, schema)
-        rescue Jing::Error => e
-          clean_abort("Jing failed with error: #{e}", doc)
-        ensure
-          f.close!
-        end
-      end
-
-      def schema_validate1(file, doc, schema)
-        file.write(to_xml(doc))
-        file.close
-        errors = Jing.new(schema, encoding: "UTF-8").validate(file.path)
-        warn "Syntax Valid!" if errors.none?
-        errors.each do |e|
-          @log.add("Metanorma XML Syntax",
-                   "XML Line #{'%06d' % e[:line]}:#{e[:column]}", e[:message])
-        end
-      end
-
-      def validate_document_fragment(xml_fragment)
-        f = add_ns_to_fragment(xml_fragment) or
-          return [true,
-                  "Fragment is not well-formed XML, not validating"]
-        begin
-          temp_schema, schema = fragment_schema(f.root.name)
-          schema or return [false, "Did not expect element #{f.root.name}"]
-          validation_errors = schema.validate(f)
-          [validation_errors.none? do |x|
-            x.to_s.include?("Did not expect element")
-          end, validation_errors]
-        ensure
-          temp_schema.unlink
-        end
-      end
-
-      def add_ns_to_fragment(xml_fragment)
-        f = Nokogiri::XML(xml_fragment, &:strict)
-        f.errors.any? || f.root.nil? and return nil
-        root_tag = f.root.name
-        f.root.namespace or
-          f = Nokogiri::XML(xml_fragment
-          .sub(/<#{root_tag}([^>]*)>/,
-               "<#{root_tag}\\1 xmlns='#{xml_namespace}'>"))
-        f
-      rescue StandardError
-        nil
-      end
-
-      def fragment_schema(root_element)
-        temp_schema = Tempfile.new(["dynamic_schema", ".rng"])
-        temp_schema.write(<<~SCHEMA)
-                  <grammar xmlns="http://relaxng.org/ns/structure/1.0">
-            <include href="#{File.join(File.dirname(__FILE__), 'isodoc-compile.rng')}">
-              <start combine="choice">
-                  <ref name="#{root_element}"/>
-              </start>
-          </include>
-                  </grammar>
-        SCHEMA
-        temp_schema.close
-        [temp_schema, Nokogiri::XML::RelaxNG(File.open(temp_schema.path))]
-      rescue StandardError # error because root_element is not in schema
-        [temp_schema, nil]
-      end
-
-      SVG_NS = "http://www.w3.org/2000/svg".freeze
-
-      WILDCARD_ATTRS =
-        "//*[@format] | //stem | //bibdata//description | " \
-        "//formattedref | //bibdata//note | //bibdata/abstract | " \
-        "//bibitem/abstract | //bibitem/note | //metanorma-extension".freeze
-
-      # RelaxNG cannot cope well with wildcard attributes. So we strip
-      # any attributes from FormattedString instances (which can contain
-      # xs:any markup, and are signalled with @format) before validation.
-      def formattedstr_strip(doc)
-        doc.xpath(WILDCARD_ATTRS, "m" => SVG_NS).each do |n|
-          n.elements.each do |e|
-            e.traverse do |e1|
-              e1.element? and e1.each { |k, _v| e1.delete(k) } # rubocop:disable Style/HashEachMethods
-            end
-          end
-        end
-        doc.xpath("//m:svg", "m" => SVG_NS).each { |n| n.replace("<svg/>") }
-        doc
       end
 
       def image_validate(doc)
